Moved function AppendH2ToNextProtos and CloneTLSConfig to internal/credentials
