+++ credentials/credentials_test.go
@@ -22,7 +22,6 @@ import (
 	"context"
 	"crypto/tls"
 	"net"
-	"reflect"
 	"strings"
 	"testing"
 
@@ -325,40 +324,4 @@ func tlsClientHandshake(conn net.Conn, _ string) (AuthInfo, error) {
 		return nil, err
 	}
 	return TLSInfo{State: clientConn.ConnectionState(), CommonAuthInfo: CommonAuthInfo{SecurityLevel: PrivacyAndIntegrity}}, nil
-}
-
-func (s) TestAppendH2ToNextProtos(t *testing.T) {
-	tests := []struct {
-		name string
-		ps   []string
-		want []string
-	}{
-		{
-			name: "empty",
-			ps:   nil,
-			want: []string{"h2"},
-		},
-		{
-			name: "only h2",
-			ps:   []string{"h2"},
-			want: []string{"h2"},
-		},
-		{
-			name: "with h2",
-			ps:   []string{"alpn", "h2"},
-			want: []string{"alpn", "h2"},
-		},
-		{
-			name: "no h2",
-			ps:   []string{"alpn"},
-			want: []string{"alpn", "h2"},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := AppendH2ToNextProtos(tt.ps); !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("AppendH2ToNextProtos() = %v, want %v", got, tt.want)
-			}
-		})
-	}
 }
